Second try

+++ PhotosService/Controllers/PhotosApiController.cs
@@ -1,16 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using PhotosService.Data;
 using PhotosService.Models;
+using PhotosService.Services;
 
 namespace PhotosService.Controllers
 {
     [ApiController]
     [Route("api/photos")]
+    [Authorize]
     public class PhotosApiController : Controller
     {
         private readonly IPhotosRepository photosRepository;
@@ -23,8 +28,11 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetPhotos(string ownerId)
+        public async Task<IActionResult> GetPhotos(string ownerId, JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != ownerId)
+                return Forbid();
+            
             var photoEntities = (await photosRepository.GetPhotosAsync(ownerId)).ToList();
             var photos = mapper.Map<List<PhotoDto>>(photoEntities);
             foreach(var photo in photos)
@@ -33,11 +41,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet("{id}/meta")]
-        public async Task<IActionResult> GetPhotoMeta(Guid id)
+        public async Task<IActionResult> GetPhotoMeta(Guid id, JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
 
             var photo = mapper.Map<PhotoDto>(photoEntity);
             photo.Url = GeneratePhotoUrl(photo);
@@ -45,8 +55,14 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet("{id}/content")]
-        public async Task<IActionResult> GetPhotoContent(Guid id)
+        public async Task<IActionResult> GetPhotoContent(Guid id, JwtSecurityToken accessToken)
         {
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+                return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
+            
             var photoContent = await photosRepository.GetPhotoContentAsync(id);
             if (photoContent == null)
                 return NotFound();
@@ -55,8 +71,11 @@ namespace PhotosService.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo)
+        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo, [FromHeader] JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != photo.OwnerId)
+                return Forbid();
+            
             var content = Convert.FromBase64String(photo.Base64Content);
             var result = await photosRepository.AddPhotoAsync(photo.Title, photo.OwnerId, content);
             if (!result)
@@ -65,11 +84,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo)
+        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo, [FromHeader] JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
 
             photoEntity.Title = photo.Title;
             var result = await photosRepository.UpdatePhotoAsync(photoEntity);
@@ -79,11 +100,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeletePhoto(Guid id)
+        public async Task<IActionResult> DeletePhoto(Guid id, JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
 
             var result = await photosRepository.DeletePhotoAsync(photoEntity);
             if (!result)
@@ -93,11 +116,40 @@ namespace PhotosService.Controllers
 
         private string GeneratePhotoUrl(PhotoDto photo)
         {
-            var relativeUrl = Url.Action(nameof(GetPhotoContent), new {
+            var relativeUrl = Url.Action(nameof(GetPhotoSignedContent), new {
                 id = photo.Id
             });
             var url = "https://localhost:6001" + relativeUrl;
-            return url;
+
+            var nowUtc = DateTime.UtcNow;
+            var signedUrl = SignedUrlHelpers.CreateSignedUrl(url, nowUtc, nowUtc.AddMinutes(5));
+            return signedUrl;
+        }
+
+        [AllowAnonymous]
+        [HttpGet("{id}/signed-content")]
+        public async Task<IActionResult> GetPhotoSignedContent(Guid id)
+        {
+            var currentUrl = HttpContext.Request.GetEncodedUrl();
+            var check = SignedUrlHelpers.CheckSignedUrl(currentUrl);
+            if (!check)
+            {
+                return Forbid();
+            }
+
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+            {
+                return NotFound();
+            }
+
+            var photoContent = await photosRepository.GetPhotoContentAsync(id);
+            if (photoContent == null)
+            {
+                return NotFound();
+            }
+
+            return File(photoContent.Content, photoContent.ContentType, photoContent.FileName);
         }
     }
 }
